kernel: i/o ports routines

+++ new/include/hal/ports64.h
@@ -8,12 +8,15 @@
 // IO Delay.
 #define io_delay() asm("out %%al,$0x80"::);
 
-unsigned char in8 (unsigned int port);
-unsigned short in16 (unsigned int port);
-unsigned int in32 (unsigned int port);
-void out8 ( unsigned int port, unsigned char data );
-void out16 (unsigned int port, unsigned short data);
-void out32 ( unsigned int port, unsigned int data );
+unsigned char  in8  (unsigned short port);
+unsigned short in16 (unsigned short port);
+unsigned int   in32 (unsigned short port);
+
+void out8  ( unsigned short port, unsigned char  data );
+void out16 ( unsigned short port, unsigned short data );
+void out32 ( unsigned short port, unsigned int   data );
+
+
 
 
 void __x86_io_delay (void);
